+++ Changelog.md
@@ -12,7 +12,8 @@ Types of changes: New, Improved, Fixed, Removed.
 - Fixed: Issue where some invalid embellishment errors would be missing the file location
 - Fixed: Zed issue with path quoting
 - Fixed: Single-note variations at the end of a bar would close after the barline
-- Improved: Added embellishment case for a grip from D to E 
+- Improved: Added embellishment case for a grip from D to E
+- Improved: Auto-detect half doublings when using `xx` after high-A
 
 ## [1.0.0]